rename datacube into MultiDataset
